+++ src/main/java/io/github/poshjosh/ratelimiter/annotation/exceptions/DuplicateNameException.java
@@ -3,7 +3,11 @@ package io.github.poshjosh.ratelimiter.annotation.exceptions;
 public class DuplicateNameException extends AnnotationProcessingException{
 
     public DuplicateNameException() {
-        super("The name of a rate related annotation may only be used once");
+        this("The id of a rate definition may only be used once");
+    }
+
+    public DuplicateNameException(String message) {
+        super(message);
     }
 
     public DuplicateNameException(String name, Object existingUse, Object attemptedUse) {